Extract common code into concerns
