+++ app/controllers/spree/adyen_redirect_controller.rb
@@ -1,25 +1,8 @@
 module Spree
   class AdyenRedirectController < AdyenController
-    before_action :restore_session, only: :confirm
-    before_action :check_signature, only: :confirm
+    include Spree::Adyen::CanConfirmPayment
 
     skip_before_action :verify_authenticity_token
-
-    # This is the entry point after an Adyen HPP payment is completed
-    def confirm
-      # Reload order as it might have changed since previously loading it
-      # from an auth notification coming in at the same time.
-      # This and the notification processing need to have a lock on the order
-      # as they both decide what to do based on whether or not the order is
-      # complete.
-      @order.with_lock do
-        if @order.complete?
-          confirm_order_already_completed
-        else
-          confirm_order_incomplete
-        end
-      end
-    end
 
     def authorise3d
       @payment = Spree::Payment.find_by(number: params[:payment_reference])
@@ -31,7 +14,7 @@ module Spree
         advance_to_confirm(@order)
         redirect_to checkout_state_path(@order.state)
       rescue Spree::Gateway::AdyenCreditCard::InvalidDetailsError
-        handle_failed_redirect
+        handle_failed_payment
       end
     end
 
@@ -46,58 +29,12 @@ module Spree
       end
     end
 
-    def handle_failed_redirect
+    def handle_failed_payment
       flash.notice = Spree.t(:payment_processing_failed)
       redirect_to checkout_state_path(@order.state)
     end
 
-    def confirm_order_incomplete
-      source = Adyen::HppSource.new(source_params)
-
-      return handle_failed_redirect unless source.authorised?
-
-      # payment is created in a 'checkout' state so that the payment method
-      # can attempt to auth it. The payment of course is already auth'd and
-      # adyen hpp's authorize implementation just returns a dummy response.
-      @order.payments.create!(
-        amount: @order.total,
-        payment_method: @payment_method,
-        source: source,
-        response_code: psp_reference,
-        state: "checkout"
-      )
-
-      if complete
-        redirect_to_order
-      else
-        #TODO void/cancel payment
-        redirect_to checkout_state_path(@order.state)
-      end
-    end
-
-    # If an authorization notification is received before the redirection the
-    # payment is created there. In this case we just need to assign the addition
-    # parameters received about the source.
-    #
-    # We do this because there is a chance that we never get redirected back
-    # so we need to make sure we complete the payment and order.
-    def confirm_order_already_completed
-      if psp_reference
-        payment = @order.payments.find_by!(response_code: psp_reference)
-      else
-        # If no psp_reference is present but the order is complete then the
-        # notification must have completed the order and created the payment.
-        # Therefore select the last Adyen payment.
-        payment =
-          @order.payments.where(source_type: "Spree::Adyen::HppSource").last
-      end
-
-      payment.source.update(source_params)
-
-      redirect_to_order
-    end
-
-    def redirect_to_order
+    def handle_successful_payment
       @current_order = nil
       flash.notice = Spree.t(:order_processed_successfully)
       flash['order_completed'] = true
@@ -106,13 +43,13 @@ module Spree
 
     def check_signature
       unless ::Adyen::HPP::Signature.verify(response_params, @payment_method.shared_secret)
-        raise "Payment Method not found."
+        raise Spree::Adyen::InvalidSignatureError, 'Signature invalid!'
       end
     end
 
     # We pass the guest token and payment method id in, pipe seperated in the
     # merchantReturnData parameter so that we can recover the session.
-    def restore_session
+    def restore_order
       guest_token, payment_method_id =
         params.fetch(:merchantReturnData).split("|")
 
@@ -123,43 +60,10 @@ module Spree
       @order = Spree::Order.find_by!(number: order_number)
     end
 
-    def source_params
-      adyen_permitted_params
-    end
-
-    def response_params
-      adyen_permitted_params
-    end
-
     # We receive `MD`, a session identifier, and `PaRes`, an
     # authentication response, from Adyen after 3d secure redirect
     def adyen_3d_params
       params.permit(:MD, :PaRes)
-    end
-
-    def adyen_permitted_params
-      params.permit(
-        :authResult,
-        :merchantReference,
-        :merchantReturnData,
-        :merchantSig,
-        :paymentMethod,
-        :pspReference,
-        :shopperLocale,
-        :skinCode)
-    end
-
-    def order_number
-      params[:merchantReference]
-    end
-
-    def psp_reference
-      params[:pspReference]
-    end
-
-    def complete
-      @order.contents.advance
-      @order.complete
     end
   end
 end
